changed names and moved to services all fo the db calls to the app and fixed imports form the diff pages
