+++ app/statistics/calculations.ts
@@ -1,6 +1,6 @@
 // calculations.ts
 import { Timestamp } from 'firebase/firestore';
-import { StatsPeriod } from './volApi';
+import { StatsPeriod } from "../../services/volunteerAnalyticsService";
 
 // Calculate date range based on period type
 export function calculateDateRange(
