+++ app/statistics/volCard.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { View, Text, TouchableOpacity } from 'react-native';
-import { VolunteerStats } from './volApi';
+import { VolunteerStats } from "../../services/volunteerAnalyticsService";
 import { Ionicons } from '@expo/vector-icons';
 
 interface VolunteerCardProps {
